TASK: Move method getAbsolutePathForPackageRessourceUri back to the ParserCache
